передача клиенту счетчика, для определения соединения

+++ simulator/udp-connection/udp-connection/include/udp-data-struct.h
@@ -104,7 +104,7 @@ struct udp_server_data_t
 {
     float           time;
 
-    quint64         msgCount;
+    int         msgCount;
 
     int             vehicleCount;
